fix: dont try delete editora used

+++ src/modules/Editora/service/DeleteEditora.service.ts
@@ -1,3 +1,5 @@
+import { ILivroRepository } from '@modules/Livro/repository/ILivroRepository.interface';
+import { EntityUsedError } from '@shared/errors/EntitieUsedError';
 import { EntityNotFoundError } from '@shared/errors/EntityNotFoundError';
 import { inject, injectable } from 'tsyringe';
 import { IEditoraRepository } from '../repository/IEditoraRepository.interface';
@@ -7,6 +9,9 @@ class DeleteEditoraService {
   constructor(
     @inject('EditoraRepository')
     private EditoraRepository: IEditoraRepository,
+
+    @inject('LivroRepository')
+    private livroRepository: ILivroRepository,
   ) {}
 
   async execute(edi_Id: string): Promise<void> {
@@ -16,6 +21,16 @@ class DeleteEditoraService {
 
     if (!editora) {
       throw new EntityNotFoundError('Editora não encontrada');
+    }
+
+    const livros = await this.livroRepository.listBy({
+      filter: {
+        liv_edi_id: editora.edi_Id,
+      },
+    });
+
+    if (livros.results.length > 0) {
+      throw new EntityUsedError('Editora sendo utilizada impossivel deletar');
     }
 
     await this.EditoraRepository.delete(editora);
